Invites (completed)
-Accepting and Declining invites work as expected
-Accept: join the group (add to group DB and remove from invite DB)
-Decline: remove from invite DB only

****NOTE: I added an additional attribute for the INVITES and GROUPS
tables called cID to represent the class ID of the group (e.g 201 from
Math 201). I saw that when I deleted an invite like "math", it would
delete everything associated with "math". So I had to differentiate
between multiple maths. So I decided to include the class ID in the
tables as well.

Test the app. I'm 95% sure this thing works.....